+++ src/common/infrastructure/interfaces/mapper.interface.ts
