+++ src/core/core.module.ts
@@ -2,7 +2,7 @@ import { Global, Logger, Module } from "@nestjs/common";
 import { APP_FILTER, APP_INTERCEPTOR } from "@nestjs/core";
 import { AppConfigService } from "@src/core/configs/app-config.service";
 import { PrismaModule } from "@src/core/database/prisma/prisma.module";
-import { AllExceptionsFilter } from "@src/core/filter/all-exceptions.filter";
+import { AllExceptionsFilter } from "@src/core/filters/all-exceptions.filter";
 import { ResponseInterceptor } from "@src/core/interceptors/response.interceptor";
 import { createWinstonLogger } from "@src/core/logger/winston.config";
 import { WinstonLogger } from "@src/core/logger/winston.logger";
