+++ src/modules/user/domain/value-objects/email.vo.ts
