+++ src/modules/user/domain/value-objects/nickname.vo.ts
