+++ src/modules/user/domain/value-objects/password.vo.ts
@@ -1,4 +1,4 @@
-import { IPasswordHasher } from "@src/modules/user/domain/interfaces/password-hasher.interface";
+import { IPasswordHasher } from "@src/modules/user/domain/service-interfaces/password-hasher.interface";
 
 export const PASSWORD_MIN_LENGTH = 8;
