+++ src/modules/user/domain/value-objects/profile-image-url.vo.ts
