+++ src/modules/user/domain/value-objects/role.vo.ts
