+++ src/modules/user/domain/value-objects/social-provider.vo.ts
