+++ src/modules/user/domain/value-objects/user-id.vo.ts
